core: first rpc_server structure
This code contains the bare bones structures for an rpc communication
with a simple frame to start hacking on the rpc mechanism. This is a much
simpler mechanism that on the seastar framework as it will allow for a callback
on the lifetime of the requests as well as stats.

I'm going to try and model this request like facebook::wangle's structure. Though
much less ambitions and more scoped. The first step is to have this skeleton

* Added rpc_server class
* Added stats class
* set app the app main template
* Added rpc-stats-printer loop

Signed-off-by: Alexander Gallego <gallego.alexx@gmail.com>