wip:adding wal_fil_manager & wal_writer_file_lease
* wal_writer_file_lease effectively takes a lease of a name
upon opening a file... it opens a DIFFERENT file. it opens it w/ an
exclusive lock, which is basically not needed at all... now that all the
writes are centralized...

so i don't need to add a prefix at all... it just needs to be
communicated w/ the file manager...

the readers can only read frofm the subscription of the file namanger...

also pending is the healing of the filesystem...

the good thing now is that this alows for a centralized point for adding
compactions, etc.