Add git configurations to ensure that git user is set
Vagrant ansible-playbook playbooks/devbox_all.yml fails because
the git module runs by default in strict mode which does not allow one
to clone repos, etc without setting user:email on gitconfig.

This patch address that. In addition, this is key for CICD env's
where it's not expected that the users have git settings set.