smf_gen: Use object API in generated code
1) Use the object API from flatbuffers now that we have a generic way of
typing the messages.

Introduce the `rpc_typed_envelope` which allows users to use a
struct-like or POD data api for their return types and lets the framework
manage the types & memory for them.
This feels more natural and is on par w/ gRPC and ApacheThrift api's.

2) rpc: add rpc_letter concepts for type conversions

rpc_letter MUST take in a root param type of a flatbuffers::Table
but will pass in as arguments flatbuffers::NativeTable via a

```

T::Pack(Builder, typename T::NativeTable) function call

```