feature: chain_replication reworked. Use all cores via map_reduce
This features introduces the notion of partition managers
per topic.

First we take the put request and consistent-hash it to cores
where each core has a manager of readers and writers
per **partition**.

This is done via xform-projection. That is we take the normal put
and split it up into writable chunks that are meant to be saved
on the filesystem. Each chunk can be concurrently saved.
Obviously we need to cap the parallelism on a file basis.

This change ONLY introduces the writers.

We also keep a cache on a rolling basis for the 'write-behind'
filesystem cache that seasar keeps around which is usually 1MB
per open writable file with direct IO

There is also a major revamp of the RPC module. Effectively
we no longer nest flatbuffers together - had some really difficlut
time getting it to work. The upside is that the metadata we manage is
now MUCH MUCh smaller - exactly 16 bytes and has room for
session management when we introduce the streaming RPC clients.

Many small scripts and utils were updated for the ./release
script to be more maintainable.