compression: add lz4 size before writing the buffer
currently lz4 either needs a very long and expensive frame with
header and footer of 20-something bytes. Or it uses the block
api wich has no frame or encoding.

The current design is to embed an int as little endian
on the first 4 bytes of the encoded buffer, so we can decode it later.

* fix rpc timeout unit test
* fix log_throw macro to use fmt::memory_writer