bugfix: include/smf/log.h macro expansion symbol collision
currently we use an fmt::MemoryWriter w; in our macros for writing
exceptional logs (LOG_THROW_IF, etc). However if the argument
passed to the function has the same name there is a symbol collision

add __smflog_<X> prefix to the variables defined in the macros to
prevent name collision.