flatbuffers: special case big buffer copies vs local buffers
flatbuffers: keep stack cache of heap ptr

Prior to this change our fbs buffer would be the
std::max(all_buffers_serialized) which would only grow and never shrink.

This patch addresses that bug. In addition, we performed several
micro benchmarks to try and stress small buffers and large buffers and
make sure that our performance was special cased for both.

Turns out that if you have small buffers, it is cheaper to memcpy
the serialized flatbuffers bulder into a perfect size destination buffer.

However, if the buffers are large we need to ensure we reset the memory
to the default of 1024 bytes.

As measured, 8 micro benchmarks this patch has:

1) No worse performance than currently for small buffers
2) No memory leaks - vs - the old way of taking the max
3) Improved performance on large buffers by not copying them