tests: Add client side timeout
There are 2 important distinctions between timeouts from
a server and client point of view. Yet we share the code.
A server is interested in a timeout *after* it parses the header

otherwise, it is happy just listening on the socket
per fiber.

On the other hand a client cares about *end-to-end* timeout.

This patch correctly handles both timeouts with a unified failure
mechanism for all using `seastar::with_timeout` and
`handle_exception`

Note that this also introduces a breaking change:
Users do not get the `make_shared` autogenerated method from the
compiler. Instead they get just a regular class that they can then
use `seastar::make_shared<ClientType>(opts..);`