cmake: only support build from source. remove cpack
The cpack is broken. It does not correctly pack all transitive deps.
It is missleading and incorrect. Removing. We can bring back this commit later
if we need it. At vectorized.io we are using a simple mustache template
+ a python wrapper which works deterministically.

We will need to open source it for smf to have correct packaging.