rpc_client: fix race condition for concurrect socket connects
With the smf::reconnect_client<T> we now have a background connect
and background socket send. We need to ensure that *creating* the socket
is exclusive with writes.

We already protected against multiple writes and protecting of sending
one payload at a time through the wire, but we did not protect
*constructing* the socket while sending data through the wire.